refactor(events): relocate arena_bridge + finalize manager.py domain-free (P4a Batch E7b)
Relocate aragora/events/arena_bridge.py -> aragora/debate/arena_bridge.py
(Relocate-UP no-shim exemption: arena_bridge's only domain coupling is a
TYPE_CHECKING-only import of DebateEvent/EventBus from aragora.debate; no
public/external consumer of the old path exists per the design doc's
consumer census). Drop the ArenaEventBridge/create_arena_bridge/
EVENT_TYPE_MAP trio from events/__init__.py and __all__ with no re-export
shim. Repoint the two runtime consumers (server/handlers/cross_pollination.py,
debate/orchestrator_memory.py) and all import/patch-path references across
4 test files (grok residual #4 census).

Finalize CrossSubscriberManager as domain-free: inline the 7 remaining
domain-free handler methods (memory_to_rlm, gauntlet_to_notification,
debate_end_to_cost_tracking, debate_end_to_explainability, culture_to_debate,
risk_warning_to_health, genesis_to_control_plane) plus CompressorProtocol
directly onto the manager class, drop the now-empty BasicHandlersMixin/
CultureHandlersMixin/StrategicHandlersMixin bases and their handlers/
subpackage (already emptied of domain-coupled reactions by E3-E6), and
update the 3 dependent test files whose regression guards imported the
deleted mixins directly.

After E7a+E7b: events -> {agents,debate,knowledge,memory,nomic,ranking,
reasoning,workflow} all cleared (only the E7a-chartered security_events
residual remains on events->debate, out of scope per orchestrator incident
#5; events->server remains for batch 2c/sweep).
